+++ src/scalecast/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.17.15'
+__version__ = '0.17.16'
 
 from .util import metrics
 import inspect